게시글,댓글 작성자 표시 및 권한, 이미지 업로드

+++ mysite/articles/views.py
@@ -6,6 +6,7 @@ from django.contrib.auth.decorators import login_required
 from .models import Article, Comment
 from .forms import ArticleForm, CommentForm
 from IPython import embed
+from django.contrib import messages
 
 # Create your views here.
 def index(request):
@@ -29,13 +30,12 @@ def detail(request, article_pk):
 
 @require_POST
 def comment_create(request, article_pk):
-    #article = Article.objects.get(pk=article_pk)
     article = get_object_or_404(Article, pk=article_pk)
     comment_form = CommentForm(request.POST)
     if comment_form.is_valid():
         comment = comment_form.save(commit=False)
         comment.article = article
-        #comment.article_id = article.pk
+        comment.user_id = request.user.pk
         comment.save()
         return redirect('articles:detail', article.pk)
     else:
@@ -54,10 +54,15 @@ def comment_delete(request, article_pk, comment_pk):
 @login_required
 def create(request):
     if request.method == "POST":
-        form = ArticleForm(request.POST)
+        form = ArticleForm(request.POST, request.FILES)
         if form.is_valid():
-            article = form.save()
+            article = form.save(commit=False)
+            article.user_id = request.user.pk
+            article.save()
+            messages.success(request, '게시글 작성 완료!')
             return redirect('articles:detail', article.pk)
+        else:
+            messages.error(request, '잘못된 데이터 입력')
     else:
         form = ArticleForm()
     context = {
@@ -68,13 +73,14 @@ def create(request):
 @login_required
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
-    if request.method == "POST":
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():
-            article = form.save()
-            return redirect('articles:detail', article.pk)
-    else:
-        form = ArticleForm(instance=article)
+    if article.user == request.user:
+        if request.method == "POST":
+            form = ArticleForm(request.POST, request.FILES, instance=article)
+            if form.is_valid():
+                article = form.save()
+                return redirect('articles:detail', article.pk)
+        else:
+            form = ArticleForm(instance=article)
     context = {
         'form': form
     }
@@ -84,5 +90,9 @@ def update(request, article_pk):
 def delete(request, article_pk):
     if request.user.is_authenticated:
         article = get_object_or_404(Article, pk=article_pk)
-        article.delete()
-    return redirect('articles:index')
+        if article.user == request.user:
+            article.delete()
+            return redirect('articles:index')
+        else:
+            return redirect('articles:detail', article.pk)
+    return redirect('articles:detail', article.pk)
